refactor(cms): renamed: extend-graphql-schema.ts -> extend-graphql-schemas/index.ts Create extend-graphql-schemas folder to locate custom GQL schemas.
